requirements.txt
opcua-asyncio
paho-mqtt
pyyaml
